separate TrimHistoryScheduler to new file
